cards number fixed

+++ src/main/java/com/mnidhub/homebanking/utils/CardUtils.java
@@ -5,10 +5,10 @@ public final class CardUtils {
     private CardUtils(){}
 
     public static String getRandomNumber(int min, int max) {
-        return ((Math.random() * (max - min)) + min ) + "-" +
-                ((Math.random() * (max - min)) + min ) + "-" +
-                ((Math.random() * (max - min)) + min ) + "-" +
-                ((Math.random() * (max - min)) + min );
+        return  (int)((Math.random() * (max - min)) + min ) + "-" +
+                (int)((Math.random() * (max - min)) + min ) + "-" +
+                (int)((Math.random() * (max - min)) + min ) + "-" +
+                (int)((Math.random() * (max - min)) + min );
     }
 
     public static int getRandomNumberCvv(int min, int max) {
